split partial_file_merge out of tagsort.cpp
